+++ CHANGELOG.md
@@ -100,6 +100,11 @@
 - **Proxy-Cache**: add `ignore_uri_case` to configuring cache-key uri to be handled as lowercase
   [#10453](https://github.com/Kong/kong/pull/10453)
 
+#### PDK
+
+- PDK now supports getting plugins' ID with `kong.plugin.get_id`.
+  [#9903](https://github.com/Kong/kong/pull/9903)
+
 ### Fixes
 
 #### Core